Move LL stuff: usb
